@@ -23,7 +23,7 @@
     "issues": "https://github.com/chesio/bc-cache/issues"
   },
   "require": {
-    "php": "^8.0",
+    "php": "^8.1",
     "composer/installers": "^1.0 || ^2.0"
   },
   "require-dev": {
@@ -31,9 +31,8 @@
     "mockery/mockery": "^1.6",
     "php-parallel-lint/php-parallel-lint": "^1.3",
     "php-stubs/wp-cli-stubs": "^2.8",
-    "phpunit/phpunit": "^9.5",
+    "phpunit/phpunit": "^10.5",
     "squizlabs/php_codesniffer": "^3.3",
-    "symfony/polyfill-php80": "^1.27",
     "szepeviktor/phpstan-wordpress": "^1.3"
   },
   "autoload-dev": {
